Add files via upload
After comparing the two sets of data, I found that conflict tends to weaken and decline with the progress of the times According to the chart, I have reason to believe that large-scale battles have begun to decline or weaken into regional violence. This is most likely due to the educational assistance of developed countries in Africa and the UN peacekeeping forces.