Add files via upload
Some speculations made from literature and data： Possible causes of nature resource conflict in Sudan（2004） Sudan’s main mineral resource reserves were found to be about 300 million tons of iron, 9 million tons of copper,oil， etc.However, the reserves of rare metals such as gold are not mentioned in various articles.During the investigation, we found that the recent conflict in Sudan has dropped significantly. We have reason to suspect that this is because the resource conflict has shifted from the past oil and ordinary ore resources to the competition for rare metals.